pyproject.toml
[tool.poetry]
name = "e-commerce"
version = "1.0"
description = "Ecommerce API."
authors = ["RezuanDzibov <rezuan.dzbov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
asgiref = "3.4.1"
certifi = "2021.10.8"
cffi = "1.15.0"
charset-normalizer = "2.0.7"
coreapi = "2.3.3"
coreschema = "0.0.4"
cryptography = "35.0.0"
defusedxml = "0.7.1"
dj-database-url = "0.5.0"
Django = "3.2.8"
django-extensions = "3.1.3"
django-filter = "21.1"
django-js-asset = "1.2.2"
django-mptt = "0.13.4"
django-phonenumber-field = "5.2.0"
django-rest-framework = "0.1.0"
django-templated-mail = "1.1.1"
djangorestframework = "3.12.4"
djangorestframework-simplejwt = "4.8.0"
djoser = "2.1.0"
drf-yasg = "1.20.0"
flake8 = "4.0.1"
gunicorn = "20.1.0"
idna = "3.3"
inflection = "0.5.1"
itypes = "1.2.0"
Jinja2 = "3.0.2"
MarkupSafe = "2.0.1"
mccabe = "0.6.1"
oauthlib = "3.1.1"
packaging = "21.0"
phonenumbers = "8.12.35"
Pillow = "8.3.2"
postgres = "4.0"
psycopg2 = "2.9.1"
psycopg2-binary = "2.9.1"
psycopg2-pool = "1.1"
pycodestyle = "2.8.0"
pycparser = "2.20"
pyflakes = "2.4.0"
PyJWT = "2.2.0"
pyparsing = "2.4.7"
python-dotenv = "0.19.1"
python3-openid = "3.2.0"
pytz = "2021.3"
requests = "2.26.0"
requests-oauthlib = "1.3.0"
"ruamel.yaml" = "0.17.16"
"ruamel.yaml.clib" = "0.2.6"
six = "1.16.0"
social-auth-app-django = "4.0.0"
social-auth-core = "4.1.0"
sqlparse = "0.4.2"
uritemplate = "4.1.1"
urllib3 = "1.26.7"
pytest = "^6.2.5"
pytest-django = "^4.4.0"
model-bakery = "^1.3.3"
factory-boy = "^3.2.0"
ipython = "^7.28.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
